2D Movement System
Inertia There is now accleratation and deacceleration on the movement of the player.
Adjustable paramaters in the inspector:
Move speed
Acceleration
Deceleration

I added jumping aswell, how ever i compleatly ducked it when it comes to showing the jumping. I used the z axis of position which is wrong, il fix this at a later date. Jumping is tied to a bool to check if the player is jumping. The idea is to have boolean as a true or false and if the player is on the false boolean they would fall down what ever they were jumping over.

Dodge roll has been added. It lockes the player in the direction they where last going during the roll, enabling an I frame. Then allows the player to move about normaly after wards.
Adjustable Parameter's in the inspector:
Speed of roll
Duration of roll
Cooldown in between rolls
Input Key
There are internal state tracking for debugging, for this as well. IE Numbers visible and true or false state ments to see if something is wrong or unexpected.